requirements.txt
# Core Flask dependencies
flask==3.0.0
python-dotenv==1.0.1
requests==2.32.3

# Data processing
feedparser==6.0.11
beautifulsoup4==4.13.5
markdown==3.5.1

# LangChain ecosystem
langchain-core==0.3.74
langchain-community==0.3.27
langchain-openai==0.3.31
langchain-huggingface==0.3.1

# Vector store and embeddings
sentence-transformers==2.2.2
chromadb==0.5.0

# Note: gunicorn not needed for Vercel deployment